+++ packages/core/src/definitions/styles.ts
@@ -111,7 +111,7 @@ export const builtInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The border of the section',
-    defaultValue: '1px solid rgba(0, 0, 0, 0)',
+    defaultValue: '0px solid rgba(0, 0, 0, 0)',
   },
   cfGap: {
     displayName: 'Gap',
@@ -371,13 +371,6 @@ export const containerBuiltInStyles: Partial<
     description: 'The max-width of the section',
     defaultValue: '1192px',
   },
-  cfMargin: {
-    displayName: 'Margin',
-    type: 'Text',
-    group: 'style',
-    description: 'The margin of the container',
-    defaultValue: '0 Auto 0 Auto',
-  },
   cfImageSizes: optionalBuiltInStyles.cfImageSizes,
 };
 
@@ -463,7 +456,7 @@ export const singleColumnBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The border of the column',
-    defaultValue: '1px solid rgba(0, 0, 0, 0)',
+    defaultValue: '0px solid rgba(0, 0, 0, 0)',
   },
   cfGap: {
     displayName: 'Gap',
@@ -529,7 +522,7 @@ export const columnsBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The margin of the columns',
-    defaultValue: '0 Auto 0 Auto',
+    defaultValue: '0 0 0 0',
   },
   cfWidth: {
     displayName: 'Width',
@@ -564,7 +557,7 @@ export const columnsBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The border of the columns',
-    defaultValue: '1px solid rgba(0, 0, 0, 0)',
+    defaultValue: '0px solid rgba(0, 0, 0, 0)',
   },
   cfBackgroundImageUrl: {
     displayName: 'Background image',