+++ packages/core/src/utils/styleUtils/styleTransformers.ts
@@ -150,7 +150,6 @@ export const transformBackgroundImage = (
     backgroundSize: matchBackgroundSize(cfBackgroundImageScaling),
   };
 };
-
 export const transformWidthSizing = ({
   value,
   cfMargin,