+++ packages/core/src/utils/utils.ts
@@ -1,9 +1,9 @@
 import {
-  CompositionTree,
-  CompositionComponentNode,
+  ExperienceTree,
+  ExperienceTreeNode,
   StyleProps,
-  CompositionDataSource,
-  CompositionUnboundValues,
+  ExperienceDataSource,
+  ExperienceUnboundValues,
   ExperienceEntry,
   ComponentDefinition,
 } from '@/types';
@@ -11,13 +11,13 @@ import { Entry } from 'contentful';
 import { ASSEMBLY_DEFAULT_CATEGORY } from '@/constants';
 
 export const getDataFromTree = (
-  tree: CompositionTree,
+  tree: ExperienceTree,
 ): {
-  dataSource: CompositionDataSource;
-  unboundValues: CompositionUnboundValues;
+  dataSource: ExperienceDataSource;
+  unboundValues: ExperienceUnboundValues;
 } => {
-  let dataSource: CompositionDataSource = {};
-  let unboundValues: CompositionUnboundValues = {};
+  let dataSource: ExperienceDataSource = {};
+  let unboundValues: ExperienceUnboundValues = {};
   const queue = [...tree.root.children];
 
   while (queue.length) {
@@ -41,8 +41,8 @@ export const getDataFromTree = (
 };
 
 type GetInsertionDataParams = {
-  dropReceiverNode: CompositionComponentNode;
-  dropReceiverParentNode: CompositionComponentNode;
+  dropReceiverNode: ExperienceTreeNode;
+  dropReceiverParentNode: ExperienceTreeNode;
   flexDirection?: StyleProps['cfFlexDirection'];
   isMouseAtTopBorder: boolean;
   isMouseAtBottomBorder: boolean;
@@ -53,7 +53,7 @@ type GetInsertionDataParams = {
 };
 
 type InsertionData = {
-  node: CompositionComponentNode;
+  node: ExperienceTreeNode;
   index: number;
 };