+++ packages/visual-editor/src/utils/assemblyUtils.spec.ts
@@ -10,7 +10,7 @@ import {
   ASSEMBLY_BLOCK_NODE_TYPE,
   ASSEMBLY_NODE_TYPE,
 } from '@contentful/experiences-core/constants';
-import type { CompositionComponentNode, CompositionNode } from '@contentful/experiences-core/types';
+import type { ExperienceTreeNode, ComponentTreeNode } from '@contentful/experiences-core/types';
 import { EditorModeEntityStore } from '@contentful/experiences-core';
 import { deserializeAssemblyNode, resolveAssembly } from './assemblyUtils';
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
@@ -32,11 +32,11 @@ describe('deserializeAssemblyNode', () => {
     assembliesRegistry.clear();
   });
 
-  it('should correctly deserialize a simple CompositionNode with no variables or children', () => {
-    const node: CompositionNode = {
+  it('should correctly deserialize a simple ComponentTreeNode with no variables or children', () => {
+    const node: ComponentTreeNode = {
       definitionId: 'assembly-id',
       variables: {},
-      children: assemblyEntry.fields.componentTree.children as CompositionNode['children'],
+      children: assemblyEntry.fields.componentTree.children as ComponentTreeNode['children'],
     };
 
     const result = deserializeAssemblyNode({
@@ -135,7 +135,7 @@ describe('resolveAssembly', () => {
   });
 
   it('should return the input node when its type is not an assembly node type', () => {
-    const node: CompositionComponentNode = {
+    const node: ExperienceTreeNode = {
       type: 'block',
       data: {
         id: '1',
@@ -162,7 +162,7 @@ describe('resolveAssembly', () => {
   });
 
   it('should return the input node when assembliesRegistry does not have the componentId', () => {
-    const node: CompositionComponentNode = {
+    const node: ExperienceTreeNode = {
       type: ASSEMBLY_NODE_TYPE,
       data: {
         blockId: 'assemblyId',
@@ -184,7 +184,7 @@ describe('resolveAssembly', () => {
   });
 
   it('should return the input node when entityStore does not have componentFields', () => {
-    const node: CompositionComponentNode = {
+    const node: ExperienceTreeNode = {
       type: ASSEMBLY_NODE_TYPE,
       data: {
         assembly: {
@@ -213,7 +213,7 @@ describe('resolveAssembly', () => {
   });
 
   it('should return the input node when entityStore is null', () => {
-    const node: CompositionComponentNode = {
+    const node: ExperienceTreeNode = {
       type: ASSEMBLY_NODE_TYPE,
       data: {
         assembly: {