+++ packages/visual-editor/src/utils/getUnboundValues.ts
@@ -1,8 +1,5 @@
 import { get } from 'lodash-es';
-import type {
-  CompositionUnboundValues,
-  CompositionVariableValueType,
-} from '@contentful/experiences-core/types';
+import type { ExperienceUnboundValues, PrimitiveValue } from '@contentful/experiences-core/types';
 
 export const getUnboundValues = ({
   key,
@@ -10,10 +7,10 @@ export const getUnboundValues = ({
   unboundValues,
 }: {
   key: string;
-  fallback: CompositionVariableValueType;
-  unboundValues: CompositionUnboundValues;
-}): CompositionVariableValueType => {
+  fallback: PrimitiveValue;
+  unboundValues: ExperienceUnboundValues;
+}): PrimitiveValue => {
   const lodashPath = `${key}.value`;
 
-  return get(unboundValues, lodashPath, fallback) as CompositionVariableValueType;
+  return get(unboundValues, lodashPath, fallback) as PrimitiveValue;
 };