+++ packages/visual-editor/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-alpha.5](https://github.com/contentful/experience-builder/compare/@contentful/experiences-visual-editor-react@0.0.1-alpha.4...@contentful/experiences-visual-editor-react@0.0.1-alpha.5) (2024-03-13)
+
+### Bug Fixes
+
+- **canvas-interactions:** selecting a component not working in nextjs app-router ([#481](https://github.com/contentful/experience-builder/issues/481)) ([a44a6ae](https://github.com/contentful/experience-builder/commit/a44a6ae44efc956f5df03247240ab1b63f5a8aa7))
+- hover outlines and default border style [ALT-323] ([#476](https://github.com/contentful/experience-builder/issues/476)) ([0378c6d](https://github.com/contentful/experience-builder/commit/0378c6d7499e9bdea6bedd632a9444dcde712984))
+
+## [0.0.1-alpha.4](https://github.com/contentful/experience-builder/compare/@contentful/experiences-visual-editor-react@0.0.1-alpha.3...@contentful/experiences-visual-editor-react@0.0.1-alpha.4) (2024-03-12)
+
+### Bug Fixes
+
+- empty state canvas now says add components to begin ([14a39d8](https://github.com/contentful/experience-builder/commit/14a39d82b20c19b8a4bc5466dd74675968930288))
+
 ## [0.0.1-alpha.3](https://github.com/contentful/experience-builder/compare/@contentful/experiences-visual-editor-react@0.0.1-alpha.2...@contentful/experiences-visual-editor-react@0.0.1-alpha.3) (2024-03-08)
 
 ### Bug Fixes